requirements.txt
python-osc==1.8.1
pyinstaller==6.3.0
ffmpeg-python==0.2.0 